requirements.txt

streamlit>=1.28.0
opencv-python-headless==4.8.0.76 
deepface>=0.0.79
pandas>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorflow>=2.12.0
torch>=2.0.0
transformers>=4.30.0
geopy>=2.3.0
